+++ backend/src/main/java/br/com/lolstats/entities/Role.java
@@ -12,24 +12,24 @@ public class Role {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	private String lane;
 
 	public Role() {
 	}
 
-	public Role(Integer id, String lane) {
+	public Role(Long id, String lane) {
 		super();
 		this.id = id;
 		this.lane = lane;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
